+++ apps/chat/src/constants/talkTo.ts
@@ -1,0 +1,4 @@
+export const SuggestedCard = {
+  id: 'suggested',
+  reference: 'suggested',
+};
