+++ apps/chat/src/types/talkTo.ts
@@ -1,0 +1,7 @@
+import { SuggestedCard } from '@/src/constants/talkTo';
+
+import { DialAIEntityModel } from './models';
+
+type SuggestedCardType = typeof SuggestedCard;
+
+export type CardType = DialAIEntityModel | SuggestedCardType;
